Add a basic template
This commit updates the blank project to include a shell of an app. Currently, all it does is attempt to access the external storage, and provide a directory listing of the files there. Selecting a subdirectory starts a new activity focusing in that directory. Selecting a file toasts the file name. Files are sorted alphabetically, and the sorting can be reversed per-activity.

No settings are present. There is no ability to read or write files themselves, or to add new directories. Additionally, the initial file scan is done on the main thread without optimisation, so can end up blocking the UI thread for folders with many children.